Move test helper functions into util module
